Add graphql properties to model classes
Add a graphql_id property to the models which are registered to the
GraphQL schema. The property can be used to get a model instance's
Global ID as seen on the GraphQL API.

Also add a graphql_node_name field to the models, as it's required by
the graphql_id property.